# openapi.yaml
---
openapi: 3.1.0
tags:
- name: Customer Service API
  description: This is a REST service to manager customer information
components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
paths:
  /customers:
    get:
      summary: "List customers, optionally filtered by first name"
      tags:
      - Customer Service API
      parameters:
      - name: firstname
        in: query
        schema:
          type: string
      responses:
        "200":
          description: Customers found
    post:
      summary: Create a new customer
      tags:
      - Customer Service API
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Customer"
        required: true
      responses:
        "201":
          description: Customer created
        "422":
          description: "Invalid customer payload supplied: id was invalidly set"
        "417":
          description: Customer could not be created
  /customers/{id}:
    put:
      summary: Update a customer by its id
      tags:
      - Customer Service API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Customer"
        required: true
      responses:
        "204":
          description: Customer updated
        "404":
          description: Customer not found
    get:
      summary: Get a customer by its id
      tags:
      - Customer Service API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Customer found
        "404":
          description: Customer not found
    delete:
      summary: Delete a customer by its id
      tags:
      - Customer Service API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Customer deleted
        "404":
          description: Customer not found
info:
  title: customer-service API
  version: 1.0.0